Добавление ветки add-remove-endpoint

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -83,6 +83,9 @@ public class FilmService {
     }
 
     public void deleteById(long filmId) {
+        if (!filmStorage.existsById(filmId)) {
+            throw new IllegalArgumentException("Фильм с id " + filmId + " не найден.");
+        }
         filmStorage.deleteById(filmId);
     }
 }
